Consolidate the stateless publish API onto PR1 registry storage primitives.
The first pass duplicated publish store/index helpers and server DTO layers across many files. This rework folds declaration handling, index checks, and the publish service into shared appregistry modules, extends RegistryObjectStore with promote support, and keeps handlers as thin auth/decode/call/encode wrappers while preserving GCS safety, signed upload headers, concurrency/idempotency, and audit behavior.
